I made the file print out all the tokens. Lots of work. The next step is to make it print out the string too (like string name or identifier).
Lesson learned: vim's tab looks like 8 spaces. I have to edit Makefile in
Sublime text in order to add tabs and that's not cool.

Having done the lexer in a switch manner, I realized it's way more efficient
to scan with a finite state machine. The code might probably be shorter and
it would be able to support more changes to the language.

Not sure how to implement error catching :/